requirements.txt
asgiref==3.8.1
contourpy==1.3.0
cycler==0.12.1
distlib==0.3.8
Django==5.1.1
django-filter==24.3
django-jalali==7.1.0
djangorestframework==3.15.2
filelock==3.16.1
fonttools==4.53.1
jalali_core==1.0.0
jdatetime==5.0.0
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.1
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
pyparsing==3.1.4
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
virtualenv==20.26.5
